+++ src/app/core/desktop.ts
@@ -1,0 +1,9 @@
+import { Injectable } from '@angular/core';
+import { remote } from 'electron';
+
+@Injectable()
+export class Desktop {
+    public openLink(url: string): void {
+        remote.shell.openExternal(url);
+    }
+}