+++ src/app/core/github-api.ts
@@ -1,17 +1,27 @@
+import { HttpClient } from '@angular/common/http';
 import { Injectable } from '@angular/core';
-import { HttpClient, HttpHeaders, HttpParams } from '@angular/common/http';
+import { Strings } from './strings';
 
-@Injectable({
-    providedIn: 'root',
-})
+@Injectable()
 export class GitHubApi {
-    constructor(private httpClient: HttpClient) {
-    }
+    constructor(private httpClient: HttpClient) {}
 
-    public async getLastestReleaseAsync(owner: string, repo: string): Promise<string> {
-        const url: string = `https://api.github.com/repos/${owner}/${repo}/releases/latest`;
-        const latestReleaseResponse: any = await this.httpClient.get<any>(url).toPromise();
+    public async getLatestReleaseAsync(owner: string, repo: string, includePrereleases: boolean): Promise<string> {
+        const url: string = `https://api.github.com/repos/${owner}/${repo}/releases`;
+        const releasesResponse: any = await this.httpClient.get<any>(url).toPromise();
 
-        return latestReleaseResponse.tag_name.replace('v', '');
+        let latestRelease: any = releasesResponse.find((x) => x.prerelease);
+
+        if (includePrereleases) {
+            latestRelease = releasesResponse.find((x) => x.prerelease);
+        } else {
+            latestRelease = releasesResponse.find((x) => !x.prerelease);
+        }
+
+        if (latestRelease != undefined && latestRelease.tag_name != undefined) {
+            return Strings.replaceFirst(latestRelease.tag_name, 'v', '');
+        }
+
+        return '';
     }
 }