+++ src/app/core/strings.ts
@@ -1,0 +1,5 @@
+export class Strings {
+    public static replaceFirst(sourceString: string, oldValue: string, newValue: string): string {
+        return sourceString.replace(oldValue, newValue);
+    }
+}